# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost'
paths:
...
  '/api/cart/book/{book}':
    post:
      summary: ''
      operationId: postApiCartBookBook
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    delete:
      summary: ''
      operationId: deleteApiCartBookBook
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: book
        description: 'The book.'
        example: 11
        required: true
        schema:
          type: integer
...
  '/api/dashboard/books/{id}':
    get:
      summary: ''
      operationId: getApiDashboardBooksId
      description: ''
      parameters: []
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - Endpoints
      security: []
    put:
      summary: ''
      operationId: putApiDashboardBooksId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: ''
                  example: voluptatibus
                  nullable: false
                author:
                  type: string
                  description: ''
                  example: officia
                  nullable: false
                publisher:
                  type: string
                  description: ''
                  example: expedita
                  nullable: false
                publicationYear:
                  type: string
                  description: ''
                  example: numquam
                  nullable: false
                lang:
                  type: string
                  description: ''
                  example: aperiam
                  nullable: false
                edition:
                  type: string
                  description: ''
                  example: libero
                  nullable: false
                translator:
                  type: string
                  description: ''
                  example: aut
                  nullable: false
                isbn:
                  type: string
                  description: ''
                  example: eius
                  nullable: false
                description:
                  type: string
                  description: ''
                  example: null
                  nullable: true
                cat_id:
                  type: string
                  description: ''
                  example: sit
                  nullable: false
                dep_id:
                  type: string
                  description: ''
                  example: odio
                  nullable: false
                sec_id:
                  type: string
                  description: ''
                  example: est
                  nullable: false
                format:
                  type: string
                  description: ''
                  example: labore
                  nullable: false
                barrow:
                  type: string
                  description: ''
                  example: enim
                  nullable: false
              required:
                - title
                - author
                - publisher
                - publicationYear
                - lang
                - edition
                - translator
                - isbn
                - cat_id
                - dep_id
                - sec_id
                - format
                - barrow
      security: []
    delete:
      summary: ''
      operationId: deleteApiDashboardBooksId
      description: ''
      parameters: []
      responses: {  }
      tags:
        - Endpoints
      security: []
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the book.'
        example: 17
        required: true
        schema:
          type: integer
...